Split off job configurator
