+++ aws/src/main/kotlin/edu/wpi/axon/aws/db/DynamoJobDB.kt
@@ -89,7 +89,7 @@ class DynamoJobDB(
 
     override fun getJobsWithStatus(status: TrainingScriptProgress): IO<List<Job>> = TODO()
 
-    override fun deleteTable(): IO<Unit> = dbClient.flatMap { dbClient ->
+    fun deleteTable(): IO<Unit> = dbClient.flatMap { dbClient ->
         IO {
             dbClient.deleteTable {
                 it.tableName(tableName)
