+++ aws/src/test/kotlin/edu/wpi/axon/aws/db/DefaultJobDBTest.kt
@@ -1,7 +1,5 @@
 package edu.wpi.axon.aws.db
 
-import com.amazonaws.regions.Regions
-import edu.wpi.axon.testutil.KoinTestFixture
 import org.jetbrains.exposed.dao.EntityID
 import org.jetbrains.exposed.dao.IntEntity
 import org.jetbrains.exposed.dao.IntEntityClass
@@ -11,11 +9,7 @@ import org.jetbrains.exposed.sql.SchemaUtils
 import org.jetbrains.exposed.sql.StdOutSqlLogger
 import org.jetbrains.exposed.sql.addLogger
 import org.jetbrains.exposed.sql.transactions.transaction
-import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
-import org.koin.core.context.startKoin
-import org.koin.dsl.module
-import software.amazon.awssdk.regions.Region
 
 object Users : IntIdTable() {
     val name = varchar("name", 50).index()
@@ -42,26 +36,15 @@ class City(id: EntityID<Int>) : IntEntity(id) {
     val users by User referrersOn Users.city
 }
 
-internal class RDSJobDBTest : KoinTestFixture() {
+internal class DefaultJobDBTest {
 
     @Test
-    @Disabled("Needs RDS supervision.")
-    fun `test rds`() {
-        startKoin {
-            modules(module {
-                single { Regions.US_EAST_1 }
-                single { Region.US_EAST_1 }
-            })
-        }
-
-        val db = RDSJobDB()
-        db.ensureConfiguration().unsafeRunSync()
-
-        Database.connect(
-            url = "jdbc:mysql://axon-cluster.cluster-cnohjc8tu8oj.us-east-1.rds.amazonaws.com:3306",
-            driver = "com.mysql.jdbc.Driver",
-            user = "axonuser",
-            password = "axonpassword"
+    fun `test creating job`() {
+        val db = DefaultJobDB(
+            Database.connect(
+                url = "jdbc:h2:mem:test",
+                driver = "org.h2.Driver"
+            )
         )
 
         transaction {
